Breakpoint.cpp was added to prevent issues from having the definition in the header file being included into multiple files.
Improved the logging system to properly print the time each message occurred in the log file, as well as prepared for an expected move to have the log system be thread-safe.

Added TextureManager, which will store any and all ShaderResourceViews an Textures (both 2D and 3D) for the duration of the program so that both ends of the resource can be properly released from DirectX. Some sort of cache clearing should be used in the future.

The FileIO_Textures files now have a LoadTexture2D function, which will attempt to load a 2D texture as a DDS file, and if that fails, it will attempt to load the file as a WIC texture, and create a ShaderResourceView and Texture2D for it that way. DDS is the preferred way, as it is more efficient (at least in code that I have to write ;D). The function is also heavily error-checked, and should be pretty stable.

RenderShape now uses the TextureManager to get a texture, rather than loading it in itself. It also no longer releases the texture, as that will be handled by the TextureManager's destructor.

The SystemInitializer now initializes and destroys the TextureManager.

Signed-off-by: Drew Ritter <dritter@razelon.com>