Ignored *.user files.
Moved all files to the Hourglass namespace instead of LevelEditor.

Create the MenuButton control which has a context menu that will appear if you press the button in a certain area (and it has a menu) with more options on what to do. This button will be used to define how to create an object.

Signed-off-by: Drew Ritter <dritter@razelon.com>