Improved adding a texture by checking if the texture we receive from the TextureManager is the same as the one already set in the requested spot. If they're different, the shape as it was is removed from the renderset, ensuring that its data is properly removed.
Buffer data now appears to properly be going to different instances of objects. The issue was that due to the way a rendering context compares textures, removing a copy of a mesh with the same shaders, but no textures, would remove data bound by an identical mesh, with the same shaders, but that had textures attached to it. The current solution was to not include the check for if "other.mTextures[i] == nullptr", as this was the cause of the issue, because of the order of comparison between contexts. This fixed the issue, but I feel like a better solution is possible.

Additional fixes include that the simulated instance ID is now properly stored as a uint instead of a float, which caused a lot of shenanigans in the shaders.

The Cut function now removes indices from the ID map if their buffer is nullptr. This should, however, be changed to remove them regardless, because those indices are no longer valid and are just bloating the map.

Fixed a pretty big issue that left destroyed shapes in the render set. I simply forget to put DESTROY_NODE in the TransparentMeshComponent's RemoveShape overload.

Signed-off-by: Drew Ritter <dritter@razelon.com>