Added Added PSD for forcefield texture.
Updated Hourglass to have a checkbox for whether or not an object should be added to the timeline. Any object with a collider that is movable will be added to the timeline automatically, but marking the Record In Timeline checkbox will also guarantee the object is recorded.

Updated Hourglass to support setting transparency for mesh components.

Updated the engine to support the new hourglass controls.

Signed-off-by: Drew Ritter <dritter@razelon.com>