Hourglass now supports setting the shaders for a mesh.
Mesh components have been rearranged to better support the system setup to have component inheritance.

Chrono::Loop now checks the binary file version in order to not incorrectly read data.

Added a new constructor to MeshComponent that allows us to set it's shaders. This is also available in one convenient function as well.

Signed-off-by: Drew Ritter <dritter@razelon.com>