Blurring is calculated on a guassian curve now. The BlurTextures function still only works with a single texture to blur, but other than that, it appears to be working perfectly.
Currently, there are controls included in the renderer to control the blur. The division button on the numpad will toggle the blur effect, which defaults to disabled. The addition and subtraction keys on the numpad will increase/decrease the guassian sigma value with delta time while the buttons are pressed. When shift is held down, the addition and subtraction keys on the numpad instead affect the downsample amount for the texture.

Signed-off-by: Drew Ritter <dritter@razelon.com>