Improved Hourglass' CPU usage by removing an unused timer.
Added specular maps to the texture mesh component, and incremented the file writer version to 6.

When opening or saving a file, the current filename will be stored and will allow Save to save to that file without a filepicker.

The Save file picker will now open to the project's resources directory, instead of Hourglass' binary directory.

Signed-off-by: Drew Ritter <dritter@razelon.com>