The Renderer now deletes the DeferredCombiner render shape.
Particles are now somewhat clearing particles better. That system is so spaghetti it's hard to tell what's right an wrong.

The pool should now ensure that the objects being added to it are devoid of components.

When a base object is destroyed, it will remove it's children from the level, and then put them back in the pool. I don't think this actually accounts for children more than one level deep.

Signed-off-by: Drew Ritter <dritter@razelon.com>