Removed unused vertex shaders.
Specular now works with light volumes. Normal mapping should as well, but I don't have any normal maps to test with at the moment, will get some after this commit.

Added a LightComponent.cpp, which is a subclass of MeshComponent, in order to simplify the sending of the light component's position and color in one buffer, which reduces wasteful bandwidth consumption.

Signed-off-by: Drew Ritter <dritter@razelon.com>