Fixed a few bugs involving switching levels and lights becoming misaligned with their data on the GPU:
SetAsLight created a render node, regardless of whether or not  level was loading or not.

Light components set mVisible to true when being created to signal that when the levels switch over, it should be kept visible.

The level will now also loop through light components in order to toggle their visibility when preparing the next level.

Signed-off-by: Drew Ritter <dritter@razelon.com>