Made the skyboxes in levels 4 and 5 not be affected by lighting.
Level 2 had a duplicate tile in the bottom section labeled "floor", allowing the player to teleport over nothingness.

Hopefully fixed issues with the Reset Level function not applying the level's rotation.

Hopefully fixed an issue with level 2's platform script undoing the level's rotation when the platform moves the player.

Signed-off-by: Drew Ritter <dritter@razelon.com>