Topmost objects are now properly drawn in front of transparent objects. Bloom addition has been moved to the Combiner stage, as it was causing issues with topmost objects having bloom on top, and also that it makes more sense having it at this stage.
Texture load warning that the texture can't be loaded as DDS has been replaced with a success message when a texture loads properly.

The specular map for the player's Wall-E hand has been corrected in all load level scripts to include the ".png" extension.

The specular and emissive maps for the Device have been corrected in all the load level scripts to include the ".png" extension.

Signed-off-by: Drew Ritter <dritter@razelon.com>